Unify browser.js and index.js, detect which XMLHttpRequest method to use there
