+++ package.json
@@ -3,7 +3,7 @@
   "version": "0.1.7",
   "description": "An streaming XHR abstraction that works in browsers and node.js",
   "main": "index.js",
-  "browser": "browser.js",
+  "browser": "index.js",
   "scripts": {
     "100%": "istanbul check-coverage --statements 100 --functions 100 --lines 100 --branches 100",
     "wd": "node test/index.js --wd",
@@ -43,7 +43,7 @@
     "extendible": "0.1.x",
     "hang": "1.0.x",
     "loads": "0.0.x",
-    "node-http-xhr": "1.1.x",
+    "node-http-xhr": "~1.2.1",
     "xhr-send": "1.0.x"
   },
   "devDependencies": {
